splitNode refactor; quadratic splitNode first version added
