+++ public/main.js
@@ -15,25 +15,25 @@ window.onload = function () {
 		// { rect: { x1: 30, x2: 33, y1: 1, y2: 3 } },
 		// { rect: { x1: 33, x2: 36, y1: 1, y2: 3 } },
 		//
-		// { rect: { x1: 4, x2: 7, y1: 6, y2: 9 } },
-		// { rect: { x1: 5, x2: 9, y1: 8, y2: 12 } },
-		// { rect: { x1: 6, x2: 11, y1: 5, y2: 9 } },
-		// { rect: { x1: 9, x2: 13, y1: 10, y2: 15 } },
-		// { rect: { x1: 18, x2: 24, y1: 18, y2: 21 } },
-		// { rect: { x1: 16, x2: 21, y1: 14, y2: 19 } },
-		// { rect: { x1: 19, x2: 23, y1: 5, y2: 9 } },
-		// { rect: { x1: 21, x2: 25, y1: 3, y2: 6 } }, //	1
-		// { rect: { x1: 15, x2: 19, y1: 7, y2: 10 } }, //	2
-		// { rect: { x1: 15, x2: 19, y1: 7, y2: 10 } },
-		// { rect: { x1: 18, x2: 20, y1: 9, y2: 12 } },
-		// { rect: { x1: 14, x2: 19, y1: 11, y2: 15 } },
-		// { rect: { x1: 16, x2: 19, y1: 14, y2: 17 } },
-		// { rect: { x1: 17, x2: 21, y1: 16, y2: 18 } },
+		{ rect: { x1: 4, x2: 7, y1: 6, y2: 9 } },
+		{ rect: { x1: 5, x2: 9, y1: 8, y2: 12 } },
+		{ rect: { x1: 6, x2: 11, y1: 5, y2: 9 } },
+		{ rect: { x1: 9, x2: 13, y1: 10, y2: 15 } },
+		{ rect: { x1: 18, x2: 24, y1: 18, y2: 21 } },
+		{ rect: { x1: 16, x2: 21, y1: 14, y2: 19 } },
+		{ rect: { x1: 19, x2: 23, y1: 5, y2: 9 } },
+		{ rect: { x1: 21, x2: 25, y1: 3, y2: 6 } }, //	1
+		{ rect: { x1: 15, x2: 19, y1: 7, y2: 10 } }, //	2
+		{ rect: { x1: 15, x2: 19, y1: 7, y2: 10 } },
+		{ rect: { x1: 18, x2: 20, y1: 9, y2: 12 } },
+		{ rect: { x1: 14, x2: 19, y1: 11, y2: 15 } },
+		{ rect: { x1: 16, x2: 19, y1: 14, y2: 17 } },
+		{ rect: { x1: 17, x2: 21, y1: 16, y2: 18 } },
 	];
 
 	// eslint-disable-next-line no-undef
 	window.rt = new RTreeJS.RTreeIterative({
-		M: 3,
+		M: 5,
 		// m: 2,
 		data: rawData,
 		// initialStackSize: 15,
