+++ src/utils/utils.ts
@@ -66,151 +66,6 @@ export const isDuplicate = (
 	return false;
 };
 
-export const swap = (
-	rdArr: RectData[],
-	nodeArr: Node[],
-	li: number,
-	ri: number
-): void => {
-	const temp: RectData = rdArr[li];
-	const tempPtr: Node = nodeArr[li];
-	rdArr[li] = rdArr[ri];
-	nodeArr[li] = nodeArr[ri];
-	rdArr[ri] = temp;
-	nodeArr[ri] = tempPtr;
-};
-
-export const splitNode = (
-	top: Node = {
-		size: 0,
-		keys: [],
-		pointers: [],
-	},
-	rectData: RectData,
-	rectDataPtr: Node,
-	M: number,
-	m: number
-): NodeSplitResult => {
-	const { keys: rdArr = [], pointers: nodeArr = [] } = top || {};
-
-	let lIndex = 0;
-	let rIndex = 0;
-	let min = Number.MAX_SAFE_INTEGER;
-	let max = 0;
-
-	rdArr[M] = rectData;
-	nodeArr[M] = rectDataPtr;
-	for (const [i, rd] of rdArr.entries()) {
-		if (rd.rect.x2 < min) {
-			min = rd.rect.x2;
-			lIndex = i;
-		}
-		if (rd.rect.x1 > max) {
-			max = rd.rect.x1;
-			rIndex = i;
-		}
-	}
-
-	swap(rdArr, nodeArr, 0, lIndex);
-	swap(rdArr, nodeArr, M, rIndex);
-
-	let lr: Rect = rdArr[0].rect; // left Rect
-	let rr: Rect = rdArr[M].rect; // right Rect
-
-	let li = 1;
-	let ri = M - 1;
-	let swapLeft: boolean;
-	let swapRight: boolean;
-	let ilr: Rect; // ith left Rect
-	let irr: Rect; // ith right Rect
-	let leftArea: number = getArea(lr);
-	let rightArea: number = getArea(rr);
-	let lTempLeftRect: Rect;
-	let lTempRightRect: Rect;
-	let rTempLeftRect: Rect;
-	let rTempRightRect: Rect;
-	let lTempLeftArea: number;
-	let lTempRightArea: number;
-	let rTempLeftArea: number;
-	let rTempRightArea: number;
-
-	while (li <= ri) {
-		ilr = rdArr[li].rect;
-		irr = rdArr[ri].rect;
-
-		lTempLeftRect = getCombinedRect(ilr, lr);
-		lTempLeftArea = getArea(lTempLeftRect);
-		lTempRightRect = getCombinedRect(ilr, rr);
-		lTempRightArea = getArea(lTempRightRect);
-		swapLeft = lTempLeftArea - leftArea > lTempRightArea - rightArea;
-
-		rTempLeftRect = getCombinedRect(irr, lr);
-		rTempLeftArea = getArea(rTempLeftRect);
-		rTempRightRect = getCombinedRect(irr, rr);
-		rTempRightArea = getArea(rTempRightRect);
-		swapRight = rTempRightArea - rightArea > rTempLeftArea - leftArea;
-
-		if (swapLeft && swapRight) {
-			swap(rdArr, nodeArr, li, ri);
-			lr = rTempLeftRect;
-			leftArea = rTempLeftArea;
-			rr = lTempRightRect;
-			rightArea = lTempRightArea;
-			li++;
-			ri--;
-		} else if (!swapLeft && !swapRight) {
-			lr = lTempLeftRect;
-			leftArea = lTempLeftArea;
-			rr = rTempRightRect;
-			rightArea = rTempRightArea;
-			li++;
-			ri--;
-		} else if (!swapLeft) {
-			if (lTempLeftArea > rTempLeftArea) {
-				swap(rdArr, nodeArr, li, ri);
-				lr = rTempLeftRect;
-				leftArea = rTempLeftArea;
-			} else {
-				lr = lTempLeftRect;
-				leftArea = lTempLeftArea;
-			}
-			li++;
-		} else {
-			if (rTempRightArea > lTempRightArea) {
-				swap(rdArr, nodeArr, li, ri);
-				rr = lTempRightRect;
-				rightArea = lTempRightArea;
-			} else {
-				rr = rTempRightRect;
-				rightArea = rTempRightArea;
-			}
-			ri--;
-		}
-		li++;
-		ri--;
-	}
-
-	const pivot: number = m;
-	top.size = pivot;
-
-	const rRdArr: RectData[] = new Array(M + 1); // right RectData Array
-	const rNodeArr: Node[] = new Array(M + 1); // right Node Array
-	let iter = pivot;
-	let count = 0;
-	while (iter < M + 1) {
-		rRdArr[count] = rdArr[iter];
-		rNodeArr[count] = nodeArr[iter];
-		count++;
-		iter++;
-	}
-
-	return {
-		rightRd: rRdArr,
-		rptrs: rNodeArr,
-		rightSize: count,
-	};
-};
-
 export const getPosToRemove = (
 	rdArr: RectData[],
 	size: number,
